Make SmartProperties more modular and configurable
The main change is that the SmartProperties module is now generated by a
function with the same name. This function will later be used to enable
constructing customized versions of SmartProperties that leverage a
different set of plugins than the default version.

This change is fully backwards compatible.
